:recycle: move methods where they belong
